# frontend/openapi.json
{"openapi":"3.0.2","info":{"title":"e-commerce-website","description":"e-commerce-website API","version":"0.1.0"},"paths":{"/api/v1/sign-in":{"post":{"tags":["Authentication"],"summary":"Sign In","operationId":"sign_in","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_sign_in_api_v1_sign_in_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/sign-up":{"post":{"tags":["Authentication"],"summary":"Sign Up","operationId":"sign_up","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/forgot-password":{"post":{"tags":["Authentication"],"summary":"Forgot Password","operationId":"forgot_password","parameters":[{"required":true,"schema":{"title":"Email","type":"string"},"name":"email","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/reset-password":{"put":{"tags":["Authentication"],"summary":"Reset Password","operationId":"reset_password","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPassword"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/change-password":{"put":{"tags":["Authentication"],"summary":"Change Password","operationId":"change_password","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePassword"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/user":{"get":{"tags":["User"],"summary":"Get User","operationId":"get_user","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/app__schemas__user__GetUser"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"delete":{"tags":["User"],"summary":"Delete User","operationId":"delete_user","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteUser"}}},"required":true},"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/user/role":{"get":{"tags":["User"],"summary":"Get User Role","operationId":"get_user_role","parameters":[{"required":true,"schema":{"title":"Access Token","type":"string"},"name":"access_token","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/user/shipping_address":{"get":{"tags":["User"],"summary":"Get User Shipping Address","operationId":"get_user_shipping_address","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserAddress"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"put":{"tags":["User"],"summary":"Put User Shipping Address","operationId":"put_user_shipping_address","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserAddress"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/user/balance":{"get":{"tags":["User"],"summary":"Get User Balance","operationId":"get_user_balance","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserBalance"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"put":{"tags":["User"],"summary":"Put User Balance","operationId":"put_user_balance","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PutUserBalance"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/wishlist":{"get":{"tags":["Wishlist"],"summary":"Get Wishlist","operationId":"get_wishlist","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetWishlist"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["Wishlist"],"summary":"Create Wishlist","operationId":"create_wishlist","parameters":[{"required":true,"schema":{"title":"Id","type":"string","format":"uuid"},"name":"id","in":"query"}],"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"delete":{"tags":["Wishlist"],"summary":"Delete Wishlist","operationId":"delete_wishlist","parameters":[{"required":true,"schema":{"title":"Id","type":"string","format":"uuid"},"name":"id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/image/{image_name}":{"get":{"tags":["Image"],"summary":"Get Image","operationId":"get_image","parameters":[{"required":true,"schema":{"title":"Image Name","type":"string"},"name":"image_name","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetImage"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/home/banner":{"get":{"tags":["Home"],"summary":"Get Banner","operationId":"get_banner","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetBanners"}}}}}}},"/api/v1/home/category":{"get":{"tags":["Home"],"summary":"Get Category With Image","operationId":"get_category_with_image","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetCategories"}}}}}}},"/api/v1/home/best-seller":{"get":{"tags":["Home"],"summary":"Get Best Seller","operationId":"get_best_seller","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetBestSeller"}}}}}}},"/api/v1/products":{"get":{"tags":["Product"],"summary":"Get Products","operationId":"get_products","parameters":[{"required":true,"schema":{"title":"Category","type":"array","items":{"type":"string","format":"uuid"}},"name":"category","in":"query"},{"required":false,"schema":{"title":"Page","minimum":1.0,"type":"integer","default":1},"name":"page","in":"query"},{"required":false,"schema":{"title":"Page Size","minimum":1.0,"type":"integer","default":100},"name":"page_size","in":"query"},{"required":false,"schema":{"title":"Sort By","pattern":"^(a_z|z_a)$","type":"string","default":"a_z"},"name":"sort_by","in":"query"},{"required":false,"schema":{"title":"Price","minimum":0.0,"type":"array","items":{"minimum":0.0,"type":"integer"},"default":[1,1000000]},"name":"price","in":"query"},{"required":false,"schema":{"title":"Condition","pattern":"^(new|used)$","type":"string","default":"new"},"name":"condition","in":"query"},{"required":false,"schema":{"title":"Product Name","type":"string"},"name":"product_name","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Products Api V1 Products Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Product"],"summary":"Update Product","operationId":"update_product","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProduct"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["Product"],"summary":"Create Product","operationId":"create_product","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProduct"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/products/{product_id}":{"delete":{"tags":["Product"],"summary":"Delete Product","operationId":"delete_product","parameters":[{"required":true,"schema":{"title":"Product Id","type":"string","format":"uuid"},"name":"product_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/products/{id}":{"get":{"tags":["Product"],"summary":"Get Product","operationId":"get_product","parameters":[{"required":true,"schema":{"title":"Id","type":"string","format":"uuid"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/app__schemas__product__GetProduct"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/products/search_image/upload":{"post":{"tags":["Product"],"summary":"Search Image Upload","operationId":"search_image_upload","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_search_image_upload_api_v1_products_search_image_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/products/search_image":{"post":{"tags":["Product"],"summary":"Search Image","operationId":"search_image","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_search_image_api_v1_products_search_image_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/categories":{"get":{"tags":["Category"],"summary":"Get Category","operationId":"get_category","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetCategory"}}}}}},"put":{"tags":["Category"],"summary":"Update Category","operationId":"update_category","parameters":[{"required":true,"schema":{"title":"Category Name","maxLength":100,"minLength":2,"type":"string"},"name":"category_name","in":"query"},{"required":true,"schema":{"title":"Id","type":"string","format":"uuid"},"name":"id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"post":{"tags":["Category"],"summary":"Create Category","operationId":"create_category","parameters":[{"required":true,"schema":{"title":"Category Name","maxLength":100,"minLength":2,"type":"string"},"name":"category_name","in":"query"}],"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"delete":{"tags":["Category"],"summary":"Delete Category","operationId":"delete_category","parameters":[{"required":true,"schema":{"title":"Id","type":"string","format":"uuid"},"name":"id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DefaultResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1":{"get":{"tags":["Sales"],"summary":"Get Sales","operationId":"get_sales","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetSales"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}}},"components":{"schemas":{"Banner":{"title":"Banner","required":["id","image","title"],"type":"object","properties":{"id":{"title":"Id","type":"string","format":"uuid"},"image":{"title":"Image","type":"string"},"title":{"title":"Title","type":"string"}}},"BestSeller":{"title":"BestSeller","required":["id","title","price","images"],"type":"object","properties":{"id":{"title":"Id","type":"string","format":"uuid"},"title":{"title":"Title","type":"string"},"price":{"title":"Price","type":"integer"},"images":{"title":"Images","type":"array","items":{"type":"string"}}}},"Body_search_image_api_v1_products_search_image_post":{"title":"Body_search_image_api_v1_products_search_image_post","required":["filedata"],"type":"object","properties":{"filedata":{"title":"Filedata","type":"string"}}},"Body_search_image_upload_api_v1_products_search_image_upload_post":{"title":"Body_search_image_upload_api_v1_products_search_image_upload_post","required":["file"],"type":"object","properties":{"file":{"title":"File","type":"string","format":"binary"}}},"Body_sign_in_api_v1_sign_in_post":{"title":"Body_sign_in_api_v1_sign_in_post","required":["username","password"],"type":"object","properties":{"grant_type":{"title":"Grant Type","pattern":"password","type":"string"},"username":{"title":"Username","type":"string"},"password":{"title":"Password","type":"string"},"scope":{"title":"Scope","type":"string","default":""},"client_id":{"title":"Client Id","type":"string"},"client_secret":{"title":"Client Secret","type":"string"}}},"ChangePassword":{"title":"ChangePassword","required":["old_password","new_password"],"type":"object","properties":{"old_password":{"title":"Old Password","type":"string"},"new_password":{"title":"New Password","type":"string"}}},"CreateImage":{"title":"CreateImage","required":["name","image_url"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"image_url":{"title":"Image Url","type":"string"}}},"CreateProduct":{"title":"CreateProduct","required":["title","brand","product_detail","images","price","category_id","condition"],"type":"object","properties":{"title":{"title":"Title","type":"string"},"brand":{"title":"Brand","type":"string"},"product_detail":{"title":"Product Detail","type":"string"},"images":{"title":"Images","type":"array","items":{"$ref":"#/components/schemas/CreateImage"}},"price":{"title":"Price","type":"integer"},"category_id":{"title":"Category Id","type":"string","format":"uuid"},"condition":{"title":"Condition","type":"string"}}},"DefaultResponse":{"title":"DefaultResponse","required":["message"],"type":"object","properties":{"message":{"title":"Message","type":"string"}}},"DeleteUser":{"title":"DeleteUser","required":["id"],"type":"object","properties":{"id":{"title":"Id","type":"string","format":"uuid"}}},"GetBanners":{"title":"GetBanners","required":["data"],"type":"object","properties":{"data":{"title":"Data","type":"array","items":{"$ref":"#/components/schemas/Banner"}}}},"GetBestSeller":{"title":"GetBestSeller","required":["data"],"type":"object","properties":{"data":{"title":"Data","type":"array","items":{"$ref":"#/components/schemas/BestSeller"}}}},"GetCategories":{"title":"GetCategories","required":["data"],"type":"object","properties":{"data":{"title":"Data","type":"array","items":{"$ref":"#/components/schemas/app__schemas__home__Category"}}}},"GetCategory":{"title":"GetCategory","required":["data"],"type":"object","properties":{"data":{"title":"Data","type":"array","items":{"$ref":"#/components/schemas/app__schemas__category__Category"}}}},"GetImage":{"title":"GetImage","required":["name","image_url"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"image_url":{"title":"Image Url","type":"string"}}},"GetSales":{"title":"GetSales","required":["data"],"type":"object","properties":{"data":{"$ref":"#/components/schemas/Sales"}}},"GetUserAddress":{"title":"GetUserAddress","required":["id","address_name","phone_number","address","city"],"type":"object","properties":{"id":{"title":"Id","type":"string","format":"uuid"},"address_name":{"title":"Address Name","type":"string"},"phone_number":{"title":"Phone Number","type":"string"},"address":{"title":"Address","type":"string"},"city":{"title":"City","type":"string"}}},"GetUserBalance":{"title":"GetUserBalance","required":["id","balance"],"type":"object","properties":{"id":{"title":"Id","type":"string","format":"uuid"},"balance":{"title":"Balance","type":"integer"}}},"GetWishlist":{"title":"GetWishlist","required":["data"],"type":"object","properties":{"data":{"title":"Data","type":"array","items":{"$ref":"#/components/schemas/app__schemas__wishlist__GetProduct"}}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"PutUserBalance":{"title":"PutUserBalance","required":["balance"],"type":"object","properties":{"balance":{"title":"Balance","type":"integer"}}},"ResetPassword":{"title":"ResetPassword","required":["token","password"],"type":"object","properties":{"token":{"title":"Token","type":"string"},"password":{"title":"Password","type":"string"}}},"Sales":{"title":"Sales","required":["total"],"type":"object","properties":{"total":{"title":"Total","type":"integer"}}},"UpdateImage":{"title":"UpdateImage","required":["id","name","image_url"],"type":"object","properties":{"id":{"title":"Id","type":"string","format":"uuid"},"name":{"title":"Name","type":"string"},"image_url":{"title":"Image Url","type":"string"}}},"UpdateProduct":{"title":"UpdateProduct","required":["id","title","brand","product_detail","images","price","category_id","condition"],"type":"object","properties":{"id":{"title":"Id","type":"string","format":"uuid"},"title":{"title":"Title","type":"string"},"brand":{"title":"Brand","type":"string"},"product_detail":{"title":"Product Detail","type":"string"},"images":{"title":"Images","type":"array","items":{"$ref":"#/components/schemas/UpdateImage"}},"price":{"title":"Price","type":"integer"},"category_id":{"title":"Category Id","type":"string","format":"uuid"},"condition":{"title":"Condition","type":"string"}}},"UserCreate":{"title":"UserCreate","required":["name","email","password","phone_number"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"email":{"title":"Email","type":"string"},"password":{"title":"Password","type":"string"},"phone_number":{"title":"Phone Number","type":"string"}}},"UserRead":{"title":"UserRead","required":["user_information","message","access_token","token_type"],"type":"object","properties":{"user_information":{"$ref":"#/components/schemas/app__schemas__authentication__GetUser"},"message":{"title":"Message","type":"string"},"access_token":{"title":"Access Token","type":"string"},"token_type":{"title":"Token Type","type":"string"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}},"app__schemas__authentication__GetUser":{"title":"GetUser","required":["name","email","phone_number","type"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"email":{"title":"Email","type":"string"},"phone_number":{"title":"Phone Number","type":"string"},"type":{"title":"Type","type":"string"}}},"app__schemas__category__Category":{"title":"Category","required":["id","title"],"type":"object","properties":{"id":{"title":"Id","type":"string","format":"uuid"},"title":{"title":"Title","type":"string"}}},"app__schemas__home__Category":{"title":"Category","required":["id","image","title"],"type":"object","properties":{"id":{"title":"Id","type":"string","format":"uuid"},"image":{"title":"Image","type":"string"},"title":{"title":"Title","type":"string"}}},"app__schemas__product__GetProduct":{"title":"GetProduct","required":["id","title","brand","product_detail","images_url","price","category_id","condition","size"],"type":"object","properties":{"id":{"title":"Id","type":"string","format":"uuid"},"title":{"title":"Title","type":"string"},"brand":{"title":"Brand","type":"string"},"product_detail":{"title":"Product Detail","type":"string"},"images_url":{"title":"Images Url","type":"array","items":{"type":"string"}},"price":{"title":"Price","type":"integer"},"category_id":{"title":"Category Id","type":"string","format":"uuid"},"condition":{"title":"Condition","type":"string"},"size":{"title":"Size","type":"array","items":{}}}},"app__schemas__user__GetUser":{"title":"GetUser","required":["id","name","email","phone_number","balance"],"type":"object","properties":{"id":{"title":"Id","type":"string","format":"uuid"},"name":{"title":"Name","type":"string"},"email":{"title":"Email","type":"string"},"phone_number":{"title":"Phone Number","type":"string"},"address_name":{"title":"Address Name","type":"string"},"address":{"title":"Address","type":"string"},"city":{"title":"City","type":"string"},"balance":{"title":"Balance","type":"integer"}}},"app__schemas__wishlist__GetProduct":{"title":"GetProduct","required":["id","title","price","image"],"type":"object","properties":{"id":{"title":"Id","type":"string","format":"uuid"},"title":{"title":"Title","type":"string"},"price":{"title":"Price","type":"integer"},"image":{"title":"Image","type":"string"}}}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/v1/sign-in"}}}}}}